refactor: :recycle: enhance `IJavadocDocTemplate` with generics for method documentation
Refactored the `IJavadocDocTemplate` interface to support generics, improving flexibility in handling Java method documentation. This change includes the removal of deprecated templates and the addition of new methods to better support RPC documentation generation.

BREAKING CHANGE: Existing implementations of `IJavadocDocTemplate` and `IDocBuildTemplate` need to be updated to accommodate the new method signatures and class structures.